add test

+++ tfdml/kernels/dml_diag_op.cc
@@ -76,12 +76,12 @@ class DmlDiagKernel : public DmlKernel {
     uint32_t stride_multiplier = is_64_bit_type ? 2 : 1;
 
     // Flatten the output into a vector and use strides to skip over zeros
-    uint32_t output_sizes[] = {1, 1, 1, input_shape.num_elements()};
+    uint32_t output_sizes[] = {1, 1, 1, static_cast<uint32_t>(input_shape.num_elements())};
     uint32_t output_strides[] = {
         0,
         0,
         0,
-        (input_shape.num_elements() + 1) * stride_multiplier,
+        (static_cast<uint32_t>(input_shape.num_elements()) + 1) * stride_multiplier,
     };
 
     DmlTensorInfo output;
